Repair the Travis CI
