+++ google-cloud-spanner-hibernate-testing/src/main/java/org/hibernate/userguide/model/Call.java
@@ -19,7 +19,6 @@ import javax.persistence.Table;
  */
 //tag::hql-examples-domain-model-example[]
 @Entity
-@Table(name = "phone_call")
 public class Call {
 
     @Id
