+++ google-cloud-spanner-hibernate-testing/src/test/java/org/hibernate/jpa/test/BaseEntityManagerFunctionalTestCase.java
@@ -18,8 +18,10 @@ import java.util.Map;
 import java.util.Properties;
 import javax.persistence.Entity;
 import javax.persistence.EntityManager;
+import javax.persistence.Inheritance;
 import javax.persistence.SharedCacheMode;
 import javax.persistence.ValidationMode;
+import javax.persistence.metamodel.EntityType;
 import javax.persistence.spi.PersistenceUnitTransactionType;
 
 import org.hibernate.annotations.Subselect;
@@ -33,6 +35,7 @@ import org.hibernate.jpa.HibernatePersistenceProvider;
 import org.hibernate.jpa.boot.spi.Bootstrap;
 import org.hibernate.jpa.boot.spi.PersistenceUnitDescriptor;
 import org.hibernate.metamodel.internal.MetamodelImpl;
+import org.hibernate.metamodel.model.domain.internal.EntityTypeImpl;
 import org.hibernate.persister.collection.AbstractCollectionPersister;
 import org.hibernate.testing.junit4.BaseUnitTestCase;
 import org.junit.AfterClass;
@@ -123,20 +126,37 @@ public abstract class BaseEntityManagerFunctionalTestCase extends BaseUnitTestCa
             entityManager.createNativeQuery(deleteQuery).executeUpdate();
           });
 
-      Arrays.stream(getAnnotatedClasses())
-          .filter(entity -> entity.getAnnotation(Subselect.class) == null)
-          .forEach(x -> {
-            String name = x.getAnnotation(Entity.class).name();
-            if (name == null || name.isEmpty()) {
-              name = x.getSimpleName();
-            }
-            entityManager.createQuery(getDeleteQuery(name)).executeUpdate();
-          });
+      for (EntityType entityType : entityManager.getMetamodel().getEntities()) {
+        if (shouldDelete(entityType)) {
+          entityManager.createNativeQuery(getDeleteQuery(entityType.getName())).executeUpdate();
+        }
+      }
 
       for (String extraTable : getExtraTablesToClear()) {
         entityManager.createNativeQuery(getDeleteQuery(extraTable)).executeUpdate();
       }
     });
+  }
+
+  /**
+   * Returns whether an Entity type generates a new table whose contents must be cleared.
+   */
+  private boolean shouldDelete(EntityType entityType) {
+    if (entityType.getSupertype() != null
+        && entityType.getSupertype().getJavaType().getAnnotation(Inheritance.class) != null) {
+      String inheritanceAnnotation =
+          entityType.getSupertype().getJavaType().getAnnotation(Inheritance.class).toString();
+
+      if (inheritanceAnnotation.equals("@javax.persistence.Inheritance(strategy=SINGLE_TABLE)")) {
+        return false;
+      }
+    }
+
+    if (entityType.getJavaType().getAnnotation(Subselect.class) != null) {
+      return false;
+    }
+
+    return true;
   }
 
   /** Returns a list of extra tables that need to be cleared before each test is run. */
