+++ google-cloud-spanner-hibernate-testing/src/test/resources/hibernate.properties
@@ -10,7 +10,7 @@
 hibernate.dialect com.google.cloud.spanner.hibernate.SpannerDialect
 hibernate.connection.driver_class com.google.cloud.spanner.jdbc.JdbcDriver
 hibernate.connection.url jdbc:cloudspanner:/projects/cloud-spanner-hibernate-ci/instances/test-instance/databases/test-database
-hibernate.hql.bulk_id_strategy=org.hibernate.jpa.test.ParallelInlineIdsOrClauseBulkIdStrategy
+hibernate.hql.bulk_id_strategy org.hibernate.jpa.test.NoopInlineIdsOrClauseBulkIdStrategy
 
 hibernate.show_sql=true
 hibernate.connection.pool_size 5
